+++ docs/_static/umami.js
@@ -1,0 +1,2 @@
+// This script is used to load the Umami analytics script asynchronously.
+<script defer src="https://cloud.umami.is/script.js" data-website-id="8cd88b63-caf5-4d82-98e2-8f765e390008"></script>