+++ docs/conf.py
@@ -75,3 +75,4 @@ toc_object_entries_show_parents = "hide"
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
 html_css_files = ["custom.css"]
+html_js_files = ["umami.js"]